pyproject.toml
[project]
name = "march-madness"
authors = [
    {name = "Broderick Turner", email = "broderickturner92@gmail.com"},
]
description = "Codebase for my submission to Kaggle's yearly March Machine Learning Mania competition"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
dynamic = ["version"]
dependencies = [
    "dill>=0.3",
    # "jaxlib<0.5.0",
    "jaxlib>=0.5.0",
    "numpyro>=0.17",
    "optax>=0.2",
    "polars>=1.22",
    "pydantic>=2.10",
    "python-dotenv>=1.0",
    "rich>=13.9",
    "scikit-learn>=1.6",
    "typer>=0.15",
]

[dependency-groups]
dev = [
    "ruff>=0.9.7",
]
viz = [
    "altair>=5.5",
    "arviz>=0.21",
    "matplotlib>=3.10",
    "streamlit>=1.43",
    "watchdog>=6.0",
]
debug = [
    "ipython>=8.32",
]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.ruff.lint]
#  https://docs.astral.sh/ruff/rules/
select = [
    "F",        # pyflakes
    "E",        # pycodestyle errors
    "W",        # pycodestyle warnings
    "I",        # isort
    "N",        # pep8-naming
    "UP",       # pyupgrade
    "B",        # flake8-bugbear
    "PTH",      # flake8-use-pathlib
    "FBT",      # flake8-boolean-trap
    "A",        # flake8-builtins
    "C4",       # flake8-comprehension
    "DTZ",      # flake8-datetimez
    "RET",      # flake8-return
    "PTH",      # flake8-use-pathlib
    "ASYNC",    # flake8-async
]
ignore = [
    "E501", # line too long
]
fixable = ["ALL"]

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.ruff]
line-length = 100
